Update test config, remove unnecessary application setting, fix tests
