+++ src/test/java/uk/gov/cslearning/catalogue/config/IntegrationTestConfig.java
@@ -1,12 +1,12 @@
 package uk.gov.cslearning.catalogue.config;
 
-import org.springframework.context.annotation.Configuration;
 import org.springframework.core.annotation.Order;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.stereotype.Component;
 
-@Configuration("IntegrationTestConfig")
 @Order(1)
+@Component
 public class IntegrationTestConfig extends WebSecurityConfigurerAdapter {
 
     @Override
